+++ src/interchange/src/protobuf.rs
@@ -7,101 +7,227 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-//! Protobuf source connector
-
-pub mod decode;
-
 use std::collections::HashSet;
 
-use anyhow::{bail, Result};
-use serde_protobuf::descriptor::{Descriptors, FieldDescriptor, FieldLabel, FieldType};
+use anyhow::{anyhow, bail, Context};
 
-use repr::{ColumnName, ColumnType, ScalarType};
+use protobuf::descriptor::field_descriptor_proto::Label;
+use protobuf::descriptor::FileDescriptorSet;
+use protobuf::reflect::{
+    FieldDescriptor, FileDescriptor, MessageDescriptor, ReflectFieldRef, ReflectValueRef,
+    RuntimeFieldType, RuntimeTypeBox,
+};
+use protobuf::{CodedInputStream, Message, MessageDyn};
 
-fn proto_message_name(message_name: &str) -> String {
-    // Prepend a . (following the serde-protobuf naming scheme to list root paths
-    // for packaged messages) if the message is part of a package and the user hasn't
-    // already specified a root path
-    if message_name.is_empty() || !message_name.contains('.') || message_name.starts_with('.') {
-        message_name.to_string()
-    } else {
-        format!(".{}", message_name)
-    }
+use ore::str::StrExt;
+use repr::{ColumnName, ColumnType, Datum, Row, ScalarType};
+
+/// A decoded description of the schema of a Protobuf message.
+#[derive(Debug)]
+pub struct DecodedDescriptors {
+    message_descriptor: MessageDescriptor,
+    columns: Vec<(ColumnName, ColumnType)>,
 }
 
-fn derive_scalar_type_from_proto_field<'a>(
-    seen_messages: &mut HashSet<&'a str>,
-    field: &'a FieldDescriptor,
-    descriptors: &'a Descriptors,
-) -> Result<ScalarType> {
-    let field_type = field.field_type(descriptors);
-    match field.field_label() {
-        FieldLabel::Required => bail!("Required field {} not supported", field.name()),
-        FieldLabel::Repeated => {
-            if let FieldType::Bytes = field_type {
-                bail!("Arrays or nested messages with bytes objects are not currently supported")
-            }
+impl DecodedDescriptors {
+    /// Builds a `DecodedDescriptors` from an encoded [`FileDescriptorSet`]
+    /// and the fully qualified name of a message inside that file descriptor
+    /// set.
+    pub fn from_bytes(bytes: &[u8], message_name: String) -> Result<Self, anyhow::Error> {
+        let fds =
+            FileDescriptorSet::parse_from_bytes(bytes).context("parsing file descriptor set")?;
+        let fds = FileDescriptor::new_dynamic_fds(fds.file);
+        let message_descriptor = fds
+            .iter()
+            .find_map(|fd| fd.message_by_full_name(&message_name))
+            .ok_or_else(|| {
+                anyhow!(
+                    "protobuf message {} not found in file descriptor set",
+                    message_name.quoted(),
+                )
+            })?;
+        let mut seen_messages = HashSet::new();
+        seen_messages.insert(message_descriptor.name().to_owned());
+        let mut columns = vec![];
+        for field in message_descriptor.fields() {
+            let name = ColumnName::from(field.get_name());
+            let ty = derive_column_type(&mut seen_messages, &field)?;
+            columns.push((name, ty))
         }
-        FieldLabel::Optional => (),
+        Ok(DecodedDescriptors {
+            message_descriptor,
+            columns,
+        })
     }
 
-    let typ = derive_scalar_type(seen_messages, field, descriptors)?;
-    Ok(match field.field_label() {
-        FieldLabel::Repeated => ScalarType::List {
-            element_type: Box::new(typ),
-            custom_oid: None,
-        },
-        _ => typ,
-    })
+    /// Describes the columns in the message.
+    ///
+    /// In other words, the return value describes the shape of the rows that
+    /// will be produced by a [`Decoder`] constructed from this
+    /// `DecodedDescriptors`.
+    pub fn columns(&self) -> &[(ColumnName, ColumnType)] {
+        &self.columns
+    }
 }
 
-fn derive_scalar_type<'a>(
-    seen_messages: &mut HashSet<&'a str>,
-    field: &'a FieldDescriptor,
-    descriptors: &'a Descriptors,
-) -> Result<ScalarType> {
-    Ok(match field.field_type(descriptors) {
-        FieldType::Bool => ScalarType::Bool,
-        FieldType::Int32 | FieldType::SInt32 | FieldType::SFixed32 => ScalarType::Int32,
-        FieldType::Int64 | FieldType::SInt64 | FieldType::SFixed64 => ScalarType::Int64,
-        FieldType::Enum(_) => ScalarType::String,
-        FieldType::Float => ScalarType::Float32,
-        FieldType::Double => ScalarType::Float64,
-        FieldType::UInt32 => bail!("Protobuf type \"uint32\" is not supported"),
-        FieldType::UInt64 => bail!("Protobuf type \"uint64\" is not supported"),
-        FieldType::Fixed32 => bail!("Protobuf type \"fixed32\" is not supported"),
-        FieldType::Fixed64 => bail!("Protobuf type \"fixed64\" is not supported"),
-        FieldType::String => ScalarType::String,
-        FieldType::Bytes => ScalarType::Bytes,
-        FieldType::Message(m) => {
+/// Decodes a particular Protobuf message from its wire format.
+#[derive(Debug)]
+pub struct Decoder {
+    message_descriptor: MessageDescriptor,
+    packer: Row,
+}
+
+impl Decoder {
+    /// Constructs a decoder for a particular Protobuf message.
+    pub fn new(descriptors: DecodedDescriptors) -> Self {
+        Decoder {
+            message_descriptor: descriptors.message_descriptor,
+            packer: Row::default(),
+        }
+    }
+
+    /// Decodes the encoded Protobuf message into a [`Row`].
+    pub fn decode(&mut self, bytes: &[u8]) -> Result<Option<Row>, anyhow::Error> {
+        let mut input_stream = CodedInputStream::from_bytes(bytes);
+        let mut message = self.message_descriptor.new_instance();
+        message.merge_from_dyn(&mut input_stream)?;
+        pack_message(&mut self.packer, &self.message_descriptor, &*message)?;
+        Ok(Some(self.packer.finish_and_reuse()))
+    }
+}
+
+fn derive_column_type(
+    seen_messages: &mut HashSet<String>,
+    field: &FieldDescriptor,
+) -> Result<ColumnType, anyhow::Error> {
+    match field.runtime_field_type() {
+        RuntimeFieldType::Singular(ty) => derive_inner_type(seen_messages, ty),
+        RuntimeFieldType::Repeated(ty) => {
+            let element_type = derive_inner_type(seen_messages, ty)?.scalar_type;
+            Ok(ColumnType {
+                nullable: false,
+                scalar_type: ScalarType::List {
+                    element_type: Box::new(element_type),
+                    custom_oid: None,
+                },
+            })
+        }
+        RuntimeFieldType::Map(_, _) => bail!("Protobuf map fields are not supported"),
+    }
+}
+
+fn derive_inner_type(
+    seen_messages: &mut HashSet<String>,
+    ty: RuntimeTypeBox,
+) -> Result<ColumnType, anyhow::Error> {
+    match ty {
+        RuntimeTypeBox::Bool => Ok(ScalarType::Bool.nullable(false)),
+        RuntimeTypeBox::I32 => Ok(ScalarType::Int32.nullable(false)),
+        RuntimeTypeBox::I64 => Ok(ScalarType::Int64.nullable(false)),
+        RuntimeTypeBox::U32 | RuntimeTypeBox::U64 => {
+            bail!("Protobuf unsigned integer types are not supported")
+        }
+        RuntimeTypeBox::F32 => Ok(ScalarType::Float32.nullable(false)),
+        RuntimeTypeBox::F64 => Ok(ScalarType::Float64.nullable(false)),
+        RuntimeTypeBox::String => Ok(ScalarType::String.nullable(false)),
+        RuntimeTypeBox::VecU8 => Ok(ScalarType::Bytes.nullable(false)),
+        RuntimeTypeBox::Enum(_) => Ok(ScalarType::String.nullable(false)),
+        RuntimeTypeBox::Message(m) => {
             if seen_messages.contains(m.name()) {
                 bail!("Recursive types are not supported: {}", m.name());
             }
-            seen_messages.insert(m.name());
+            seen_messages.insert(m.name().to_owned());
             let mut fields = Vec::with_capacity(m.fields().len());
             for field in m.fields() {
-                let column_name = ColumnName::from(field.name());
-                let scalar_type =
-                    derive_scalar_type_from_proto_field(seen_messages, field, descriptors)?;
-                let nullable = match field.field_label() {
-                    FieldLabel::Optional => true,
-                    FieldLabel::Repeated | FieldLabel::Required => false,
-                };
-                let column_type = ColumnType {
-                    scalar_type,
-                    nullable,
-                };
+                let column_name = ColumnName::from(field.get_name());
+                let column_type = derive_column_type(seen_messages, &field)?;
                 fields.push((column_name, column_type))
             }
             seen_messages.remove(m.name());
-            ScalarType::Record {
+            let ty = ScalarType::Record {
                 fields,
                 custom_oid: None,
                 custom_name: None,
+            };
+            Ok(ty.nullable(true))
+        }
+    }
+}
+
+fn pack_message(
+    packer: &mut Row,
+    message_desc: &MessageDescriptor,
+    message: &dyn MessageDyn,
+) -> Result<(), anyhow::Error> {
+    for field_desc in message_desc.fields() {
+        pack_field(packer, &field_desc, message)?;
+    }
+    Ok(())
+}
+
+fn pack_field(
+    packer: &mut Row,
+    field_desc: &FieldDescriptor,
+    message: &dyn MessageDyn,
+) -> Result<(), anyhow::Error> {
+    match field_desc.get_reflect(message) {
+        ReflectFieldRef::Optional(None) => {
+            if field_desc.get_proto().get_label() == Label::LABEL_REQUIRED {
+                bail!(
+                    "protobuf message missing required field {}",
+                    field_desc.get_name()
+                );
+            }
+            match field_desc.singular_runtime_type() {
+                RuntimeTypeBox::Message(_) => packer.push(Datum::Null),
+                _ => pack_value(packer, field_desc, field_desc.singular_default_value())?,
             }
         }
-        FieldType::Group => bail!("Unions are currently not supported"),
-        FieldType::UnresolvedMessage(m) => bail!("Unresolved message {} not supported", m),
-        FieldType::UnresolvedEnum(e) => bail!("Unresolved enum {} not supported", e),
-    })
+        ReflectFieldRef::Optional(Some(value)) => pack_value(packer, field_desc, value)?,
+        ReflectFieldRef::Repeated(values) => packer.push_list_with(|packer| {
+            for value in values {
+                pack_value(packer, field_desc, value)?;
+            }
+            Ok::<_, anyhow::Error>(())
+        })?,
+        ReflectFieldRef::Map(_) => {
+            bail!("internal error: unexpected map field while decoding protobuf")
+        }
+    }
+    Ok(())
+}
+
+fn pack_value(
+    packer: &mut Row,
+    field_desc: &FieldDescriptor,
+    value: ReflectValueRef,
+) -> Result<(), anyhow::Error> {
+    match value {
+        ReflectValueRef::Bool(false) => packer.push(Datum::False),
+        ReflectValueRef::Bool(true) => packer.push(Datum::True),
+        ReflectValueRef::I32(i) => packer.push(Datum::Int32(i)),
+        ReflectValueRef::I64(i) => packer.push(Datum::Int64(i)),
+        ReflectValueRef::F32(f) => packer.push(Datum::Float32(f.into())),
+        ReflectValueRef::F64(f) => packer.push(Datum::Float64(f.into())),
+        ReflectValueRef::String(s) => packer.push(Datum::String(s)),
+        ReflectValueRef::Bytes(s) => packer.push(Datum::Bytes(s)),
+        ReflectValueRef::Enum(enum_desc, i) => match enum_desc.get_value_by_number(i) {
+            None => {
+                bail!(
+                    "error decoding protobuf: enum value {} is missing while decoding field {}",
+                    i,
+                    field_desc.get_name()
+                );
+            }
+            Some(ev) => packer.push(Datum::String(ev.get_name())),
+        },
+        ReflectValueRef::Message(m) => {
+            packer.push_list_with(|packer| pack_message(packer, &m.descriptor_dyn(), &*m))?
+        }
+        ReflectValueRef::U32(_) | ReflectValueRef::U64(_) => bail!(
+            "internal error: unexpected value while decoding protobuf message: {:?}",
+            value
+        ),
+    }
+    Ok(())
 }